requirements.txt
coreapi==2.3.3
Django==4.1
djangorestframework==3.13.1
pyyaml==6.0
uritemplate==4.1.1